Add files via upload
This game will present you with a peek into a Trex's hard and dangerous life. You will have to face many obstacles, and face many enemies. One wrong step, or one decision made to late, can result you in a life full of pain and misery. Are you ready for the challenge? Click the play button to challenge life, to challenge destiny!